requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
numpy==1.26.3
torch==2.1.2
vllm==0.4.0
transformers==4.37.2
snac==0.1.2
pydantic==2.6.1
httptools==0.6.1
setuptools==69.1.1
tqdm==4.66.2
websockets==12.0
accelerate>=0.21.0
deepspeed>=0.11.0
